+++ src/passes/Intrinsics.cpp
@@ -39,9 +39,11 @@ struct IntrinsicLowering : public WalkerPass<PostWalker<IntrinsicLowering>> {
       // forgets to optimize.
       Builder builder(*getModule());
       if (auto* refFunc = target->dynCast<RefFunc>()) {
-        replaceCurrent(builder.makeCall(refFunc->func, operands, curr->type));
+        replaceCurrent(builder.makeCall(
+          refFunc->func, operands, curr->type, curr->isReturn));
       } else {
-        replaceCurrent(builder.makeCallRef(target, operands, curr->type));
+        replaceCurrent(
+          builder.makeCallRef(target, operands, curr->type, curr->isReturn));
       }
     }
   }