+++ test/binaryen.js/sourcemap.js
@@ -8,9 +8,9 @@ console.log();
 var expr = module.i32.const(1);
 var body = module.block("", [
   expr
-], binaryen.i32);
+], binaryen.Type.i32);
 
-var func = module.addFunction("main", binaryen.none, binaryen.i32, [], body);
+var func = module.addFunction("main", binaryen.Type.none, binaryen.Type.i32, [], body);
 
 module.setDebugLocation(func, expr, fileIndex, 1, 2);
 module.setDebugLocation(func, body, fileIndex, 0, 3);