+++ ts/src/classes/module/Module.ts
@@ -19,20 +19,7 @@ import {
 	type ModuleRef,
 	type SideEffect,
 	type TableRef,
-	type Type,
-	i32,
-	i64,
-	f32,
-	f64,
-	v128,
-	anyref,
-	eqref,
-	i31ref,
-	structref,
-	arrayref,
-	funcref,
-	externref,
-	stringref,
+	Type,
 } from "../../constants.ts";
 import {
 	type ExpressionBuilder,
@@ -213,19 +200,19 @@ export class Module {
 	 */
 	pop(typ: Type): ExpressionRef {
 		if ([
-			i32,
-			i64,
-			f32,
-			f64,
-			v128,
-			anyref,
-			eqref,
-			i31ref,
-			structref,
-			arrayref,
-			funcref,
-			externref,
-			stringref,
+			Type.i32,
+			Type.i64,
+			Type.f32,
+			Type.f64,
+			Type.v128,
+			Type.anyref,
+			Type.eqref,
+			Type.i31ref,
+			Type.structref,
+			Type.arrayref,
+			Type.stringref,
+			Type.funcref,
+			Type.externref,
 		].includes(typ)) {
 			return BinaryenObj["_BinaryenPop"](this[PTR], typ);
 		} else {
@@ -233,14 +220,14 @@ export class Module {
 		}
 	}
 
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly funcref = {pop: () => { BinaryenObj.printWarn("`.funcref.pop()` is deprecated; use `.pop(funcref)` instead."); return this.pop(funcref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly externref = {pop: () => { BinaryenObj.printWarn("`.externref.pop()` is deprecated; use `.pop(externref)` instead."); return this.pop(externref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly anyref = {pop: () => { BinaryenObj.printWarn("`.anyref.pop()` is deprecated; use `.pop(anyref)` instead."); return this.pop(anyref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly eqref = {pop: () => { BinaryenObj.printWarn("`.eqref.pop()` is deprecated; use `.pop(eqref)` instead."); return this.pop(eqref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly i31ref = {pop: () => { BinaryenObj.printWarn("`.i31ref.pop()` is deprecated; use `.pop(i31ref)` instead."); return this.pop(i31ref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly structref = {pop: () => { BinaryenObj.printWarn("`.structref.pop()` is deprecated; use `.pop(structref)` instead."); return this.pop(structref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly arrayref = {pop: () => { BinaryenObj.printWarn("`.arrayref.pop()` is deprecated; use `.pop(arrayref)` instead."); return this.pop(arrayref); }};
-	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly stringref = {pop: () => { BinaryenObj.printWarn("`.stringref.pop()` is deprecated; use `.pop(stringref)` instead."); return this.pop(stringref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly funcref = {pop: () => { BinaryenObj.printWarn("`.funcref.pop()` is deprecated; use `.pop(Type.funcref)` instead."); return this.pop(Type.funcref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly externref = {pop: () => { BinaryenObj.printWarn("`.externref.pop()` is deprecated; use `.pop(Type.externref)` instead."); return this.pop(Type.externref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly anyref = {pop: () => { BinaryenObj.printWarn("`.anyref.pop()` is deprecated; use `.pop(Type.anyref)` instead."); return this.pop(Type.anyref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly eqref = {pop: () => { BinaryenObj.printWarn("`.eqref.pop()` is deprecated; use `.pop(Type.eqref)` instead."); return this.pop(Type.eqref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly i31ref = {pop: () => { BinaryenObj.printWarn("`.i31ref.pop()` is deprecated; use `.pop(Type.i31ref)` instead."); return this.pop(Type.i31ref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly structref = {pop: () => { BinaryenObj.printWarn("`.structref.pop()` is deprecated; use `.pop(Type.structref)` instead."); return this.pop(Type.structref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly arrayref = {pop: () => { BinaryenObj.printWarn("`.arrayref.pop()` is deprecated; use `.pop(Type.arrayref)` instead."); return this.pop(Type.arrayref); }};
+	/** @deprecated Use {@link Module#pop} instead. @category Expression Manipulation */ readonly stringref = {pop: () => { BinaryenObj.printWarn("`.stringref.pop()` is deprecated; use `.pop(Type.stringref)` instead."); return this.pop(Type.stringref); }};
 
 	/**
 	 * Gets the side effects of the specified expression.
@@ -303,7 +290,7 @@ export class Module {
 	/** @deprecated Use {@link Module#memories | `this.memories.set`} instead. */ @replacedBy("`this.memories.set`") setMemory(initial: number, maximum: number, exportName: string, segments?: readonly any[], shared?: boolean, memory64?: boolean, internalName?: string) { return this.memories.set(initial, maximum, exportName, segments, shared, memory64, internalName); }
 	/** @deprecated Use {@link Module#memories | `this.memories.has`} instead. */ @replacedBy("`this.memories.has`") hasMemory() { return this.memories.has(); }
 
-	/** @deprecated Use {@link Module#tables | `this.tables.add`} instead. */ @replacedBy("`this.tables.add`") addTable(name: string, initial: number, maximum: number, type: Type = funcref, init?: ExpressionRef) { return this.tables.add(name, initial, maximum, type, init); }
+	/** @deprecated Use {@link Module#tables | `this.tables.add`} instead. */ @replacedBy("`this.tables.add`") addTable(name: string, initial: number, maximum: number, type: Type = Type.funcref, init?: ExpressionRef) { return this.tables.add(name, initial, maximum, type, init); }
 	/** @deprecated Use {@link Module#tables | `this.tables.get`} instead. */ @replacedBy("`this.tables.get`") getTable(name: string) { return this.tables.get(name); }
 	/** @deprecated Use {@link Module#tables | `this.tables.getByIndex`} instead. */ @replacedBy("`this.tables.getByIndex`") getTableByIndex(index: number) { return this.tables.getByIndex(index); }
 	/** @deprecated Use {@link Module#tables | `this.tables.getSegments`} instead. */ @replacedBy("`this.tables.getSegments`") getTableSegments(table: TableRef) { return this.tables.getSegments(table); }